server_app/requirements.txt
Flask==2.2.2
keras==2.11.0
mtcnn==0.1.1
numpy==1.23.5
Pillow==9.3.0
tensorflow==2.11.0
opencv-python-headless